refactor: ensure proper long type casting in various methods

+++ src/main/java/cn/nukkit/utils/VarInt.java
@@ -26,7 +26,7 @@ public final class VarInt {
      */
     public static long encodeZigZag32(int v) {
         // Note:  the right-shift must be arithmetic
-        return ((v << 1) ^ (v >> 31)) & 0xFFFFFFFFL;
+        return (((long) v << 1) ^ (v >> 31)) & 0xFFFFFFFFL;
     }
 
     /**
